refactor: remove CRA demo and change files of component to root path
