+++ .gitignore
@@ -9,9 +9,7 @@
 /coverage
 
 # production
-/build
-/radar-render/node_modules
-/radar-render/dist
+/dist
 
 # misc
 .DS_Store
