+++ package.json
@@ -1,51 +1,63 @@
 {
-  "name": "challengue",
-  "version": "0.1.0",
-  "homepage": "./",
-  "private": true,
+  "name": "radar-render",
+  "version": "2.0.7",
+  "description": "Render sections and targets on a radar",
+  "keywords": [
+    "d3js",
+    "react",
+    "render",
+    "pie",
+    "targets",
+    "radar"
+  ],
+  "main": "dist/index.cjs.js",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/PunteriaCero/torta/tree/main/radar-render"
+  },
+  "authors": [
+    "Enrique Villón <evillon@baufest.com>",
+    "Hernan Lavrencic <hlavrencic@baufest.com>",
+    "Lucio Penza <lpenza@baufest.com>",
+    "Luis Puerto <lpuerto@baufest.com>",
+    "Michael Keb <mikebizne@gmail.com>"
+  ],
+  "license": "ISC",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "start": "webpack serve --open",
+    "build": "rollup -c",
+    "build-demo": "webpack --mode production"
+  },
   "dependencies": {
+    "@babel/runtime": "^7.12.5",
+    "d3": "^7.8.5"
+  },
+  "peerDependencies": {
+    "react": ">=15.0.1",
+    "react-dom": ">=15.0.1"
+  },
+  "devDependencies": {
     "@emotion/react": "^11.11.1",
     "@emotion/styled": "^11.11.0",
     "@mui/material": "^5.14.7",
-    "@testing-library/jest-dom": "^5.17.0",
-    "@testing-library/react": "^13.4.0",
-    "@testing-library/user-event": "^13.5.0",
-    "d3": "7.8.5",
-    "radar-render": "^2.1.2",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "react-redux": "^8.1.2",
-    "react-scripts": "5.0.1",
-    "web-vitals": "^2.1.4"
-  },
-  "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject",
-    "lint": "eslint ./src/",
-    "lint:fix": "eslint ./src/ --fix"
-  },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ]
-  },
-  "devDependencies": {
-    "@babel/runtime": "7.13.8",
-    "typescript": "4.1.3"
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
+    "@babel/cli": "^7.12.10",
+    "@babel/core": "^7.12.10",
+    "@babel/plugin-transform-runtime": "^7.12.10",
+    "@babel/preset-env": "^7.12.11",
+    "@babel/preset-react": "^7.12.10",
+    "@rollup/plugin-babel": "^5.2.2",
+    "autoprefixer": "^10.4.16",
+    "babel-loader": "^9.1.3",
+    "css-loader": "^6.8.1",
+    "html-loader": "^4.2.0",
+    "html-webpack-plugin": "^5.5.3",
+    "rollup-plugin-serve": "^2.0.2",
+    "rollup-plugin-sourcemaps": "^0.6.3",
+    "rollup-plugin-styles": "^3.12.2",
+    "style-loader": "^3.3.3",
+    "webpack": "^5.89.0",
+    "webpack-cli": "^5.1.4",
+    "webpack-dev-server": "^4.15.1"
   }
 }
