--- radar-render/package.json
@@ -1,63 +1,0 @@
-{
-  "name": "radar-render",
-  "version": "2.0.7",
-  "description": "Render sections and targets on a radar",
-  "keywords": [
-    "d3js",
-    "react",
-    "render",
-    "pie",
-    "targets",
-    "radar"
-  ],
-  "main": "dist/index.cjs.js",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/PunteriaCero/torta/tree/main/radar-render"
-  },
-  "authors": [
-    "Enrique Villón <evillon@baufest.com>",
-    "Hernan Lavrencic <hlavrencic@baufest.com>",
-    "Lucio Penza <lpenza@baufest.com>",
-    "Luis Puerto <lpuerto@baufest.com>",
-    "Michael Keb <mikebizne@gmail.com>"
-  ],
-  "license": "ISC",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "webpack serve --open",
-    "build": "rollup -c",
-    "build-demo": "webpack --mode production"
-  },
-  "dependencies": {
-    "@babel/runtime": "^7.12.5",
-    "d3": "^7.8.5"
-  },
-  "peerDependencies": {
-    "react": ">=15.0.1",
-    "react-dom": ">=15.0.1"
-  },
-  "devDependencies": {
-    "@emotion/react": "^11.11.1",
-    "@emotion/styled": "^11.11.0",
-    "@mui/material": "^5.14.7",
-    "@babel/cli": "^7.12.10",
-    "@babel/core": "^7.12.10",
-    "@babel/plugin-transform-runtime": "^7.12.10",
-    "@babel/preset-env": "^7.12.11",
-    "@babel/preset-react": "^7.12.10",
-    "@rollup/plugin-babel": "^5.2.2",
-    "autoprefixer": "^10.4.16",
-    "babel-loader": "^9.1.3",
-    "css-loader": "^6.8.1",
-    "html-loader": "^4.2.0",
-    "html-webpack-plugin": "^5.5.3",
-    "rollup-plugin-serve": "^2.0.2",
-    "rollup-plugin-sourcemaps": "^0.6.3",
-    "rollup-plugin-styles": "^3.12.2",
-    "style-loader": "^3.3.3",
-    "webpack": "^5.89.0",
-    "webpack-cli": "^5.1.4",
-    "webpack-dev-server": "^4.15.1"
-  }
-}
