--- radar-render/src/styles.css
@@ -1,22 +1,0 @@
-body {
-  background-color: rgb(78, 78, 78);
- }
-
-.tableContainer {
-  background-color: rgba(0, 0, 0, 0.274);
-  display: flex;
-  flex-direction: column;
-  gap: 2.5rem;
-  justify-content: center;
-  align-items: center;
-  border-radius: 15px;
-  padding: 0.5rem;
-}
-
-circle.end-circle,
-circle.start-circle {
-  fill: rgb(255, 255, 255);
-  cursor: nwse-resize;
-  position: fixed;
-  z-index: 100 !important;
-}
