--- src/App.css
@@ -1,67 +1,0 @@
-.App {
-  text-align: center;
-  background-color: rgb(56, 56, 56);
-  min-height: 100vh;
-  display: flex;
-  align-items: center;
-  justify-content: space-around;
-
-}
-
-.container{
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  margin: 2rem;
-  background-color: rgb(63, 63, 63);
-  gap:5rem;
-  /* width: 100%; */
-  padding: 2rem;
-  border-radius: 10px;
-}
-
-.tableContainer{
-background-color:rgba(0, 0, 0, 0.274);
-display: flex;
-flex-direction: column;
-gap: 2.5rem;
-justify-content: center;
-align-items: center;
-border-radius: 15px;
-padding: 0.5rem;
-}
-
-.App-logo {
-  height: 40vmin;
-  pointer-events: none;
-}
-
-@media (prefers-reduced-motion: no-preference) {
-  .App-logo {
-    animation: App-logo-spin infinite 20s linear;
-  }
-}
-
-.App-header {
-  background-color: #282c34;
-  min-height: 100vh;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  font-size: calc(10px + 2vmin);
-  color: white;
-}
-
-.App-link {
-  color: #61dafb;
-}
-
-@keyframes App-logo-spin {
-  from {
-    transform: rotate(0deg);
-  }
-  to {
-    transform: rotate(360deg);
-  }
-}
