--- src/components/index.js
@@ -1,2 +1,0 @@
-export { default as DataTable } from './dataTable';
-export { default as Slider } from './slider';
