+++ src/data/index.js
@@ -9,7 +9,7 @@ export const data = {
       startElevation: 4,
       endElevation: 100,
       color: 'red',
-      selected: true,
+      selected: false,
       value: 10,
     },
 
@@ -22,7 +22,7 @@ export const data = {
       startElevation: 4,
       endElevation: 1,
       color: 'gray',
-      selected: false,
+      selected: true,
       value: 10,
     },
 
@@ -53,39 +53,6 @@ export const data = {
   ],
 
   targets: [
-    {
-      label: '3',
-      angle: 90,
-      radius: 0.89,
-      elevation: 5,
-      color: 'purple',
-      selected: true,
-    },
-    {
-      label: '1',
-      angle: 340,
-      radius: 0.3,
-      elevation: 5,
-      color: 'orange',
-      selected: false,
-    },
-    {
-      label: '2',
-      angle: 160,
-      radius: 0.5,
-      elevation: 5,
-      color: 'blue',
-      selected: false,
-    },
-  ],
-};
-
-export const dataDos = {
-  sections: [
-    // {label, startAngle, endAngle, innerRadius, outerRadius, startElevation, endElevation, color, selected},
-  ],
-  targets: [
-    // {label, angle, radius, elevation, color, selected},
     {
       label: '3',
       angle: 90,
