+++ src/index.js
@@ -1,17 +1,11 @@
-import React from 'react';
-import ReactDOM from 'react-dom/client';
-import './index.css';
-import App from './App';
-import reportWebVitals from './reportWebVitals';
+import RadarComponent from './components/RadarComponent.js';
+import { generateReferencesDOM, setPositionCircle } from './utils/index.js';
 
-const root = ReactDOM.createRoot(document.getElementById('root'));
-root.render(
-  <React.StrictMode>
-    <App />
-  </React.StrictMode>
-);
-
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
-reportWebVitals();
+const returnLibrary = () => {
+  return {
+    RadarComponent: RadarComponent,
+    ReferenceSectionSVG: generateReferencesDOM,
+    UpdatePositionCircle: setPositionCircle,
+  };
+};
+export default returnLibrary();
