+++ src/styles.css
@@ -1,4 +1,5 @@
 body {
+  background-color: rgb(78, 78, 78);
   margin: 0;
   font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
     'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
@@ -10,6 +11,17 @@ body {
 code {
   font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
     monospace;
+}
+
+.tableContainer {
+  background-color: rgba(0, 0, 0, 0.274);
+  display: flex;
+  flex-direction: column;
+  gap: 2.5rem;
+  justify-content: center;
+  align-items: center;
+  border-radius: 15px;
+  padding: 0.5rem;
 }
 
 circle.end-circle,
